+++ VS/horizonResource/code/source/resource/texture.cpp
@@ -8,8 +8,8 @@ using std::stringstream;
 texture::texture(){
 	UV = {{0,0},{0,0},{0,0}};
 }
-int texture::w(){if(!bmp){return 0;} return UV.size.x*bmp->width;}
-int texture::h(){if(!bmp){return 0;} return UV.size.y*bmp->height;}
+int texture::w(){if(!bmp){return 0;} return (int)(UV.size.x*bmp->width);}
+int texture::h(){if(!bmp){return 0;} return (int)(UV.size.y*bmp->height);}
 vec2 texture::size(){return {w(),h()};} 
 rect texture::getRect(){return rect(size());}
